wmbiff: FTBFS with initialize_gnutls
The first argument for the function initialize_gnutls is an intptr_t
so wmbiff fails to build from source. This patch changes the current
int to intptr_t.

+++ wmbiff/wmbiff/tlsComm.c
@@ -657,7 +657,7 @@ void handle_gnutls_read_error(int readbytes, struct connection_state *scs)
 
 #else
 /* declare stubs when tls isn't compiled in */
-struct connection_state *initialize_gnutls(UNUSED(int sd),
+struct connection_state *initialize_gnutls(UNUSED(intptr_t sd),
 										   UNUSED(char *name),
 										   UNUSED(Pop3 pc),
 										   UNUSED(const char
